+++ lawapplication/src/App.js
@@ -1,24 +1,58 @@
-import logo from './logo.svg';
-import './App.css';
+import React, { useEffect, useState } from "react";
+import Header from "./Header";
+import Search from "./Search";
+import MessageList from "./MessageList";
+import NewMessage from "./NewMessage";
+
+const testUser = { username: "Duane" };
 
 function App() {
+  const [isDarkMode, setIsDarkMode] = useState(true);
+  const [messages, setMessages] = useState([]);
+  const [search, setSearch] = useState("");
+
+  useEffect(() => {
+    fetch("http://127.0.0.1:4000/messages")
+      .then((r) => r.json())
+      .then((messages) => setMessages(messages));
+  }, []);
+
+  function handleAddMessage(newMessage) {
+    setMessages([...messages, newMessage]);
+  }
+
+  function handleDeleteMessage(id) {
+    const updatedMessages = messages.filter((message) => message.id !== id);
+    setMessages(updatedMessages);
+  }
+
+  function handleUpdateMessage(updatedMessageObj) {
+    const updatedMessages = messages.map((message) => {
+      if (message.id === updatedMessageObj.id) {
+        return updatedMessageObj;
+      } else {
+        return message;
+      }
+    });
+    setMessages(updatedMessages);
+  }
+
+  const displayedMessages = messages.filter((message) =>
+    message.body.toLowerCase().includes(search.toLowerCase())
+  );
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
+    <main className={isDarkMode ? "dark-mode" : ""}>
+      <Header isDarkMode={isDarkMode} onToggleDarkMode={setIsDarkMode} />
+      <Search search={search} onSearchChange={setSearch} />
+      <MessageList
+        messages={displayedMessages}
+        currentUser={testUser}
+        onMessageDelete={handleDeleteMessage}
+        onUpdateMessage={handleUpdateMessage}
+      />
+      <NewMessage currentUser={testUser} onAddMessage={handleAddMessage} />
+    </main>
   );
 }
